feat(sdk)!: curate exports, move test doubles to /testing, tree-shakeable entry (plan 043)
- sideEffects: false; noble sync-hash init is explicit at point of use
  (crypto/Signer, did/KeyManager) instead of a root side-effect import
- exports map curated to ., ./cel, ./testing, ./types, ./package.json;
  the 11 internal wildcards are gone
- OrdMockProvider/FeeOracleMock move to @originals/sdk/testing
- root gains Verifier, EdDSACryptosuiteManager, createDocumentLoader,
  currentControllerVm; loses retry/circuit-breaker infra exports
